Move schema files
